Consolidate some graph files.
